Create monitor zones
I have a 65inch Lg 4k tv I am using as a monitor / Tv and want to use monitor side when on windows 10 to trade currencies. I also have a couple smaller monitors. My issue is that I like the way the screen looks on 4k unit . On using power toys to have multiple window zones I like it BUT I cant find a way to make zone size proportionate to the window, including font size. Any suggestions as I want to have 10-12 small chart zones on LG screen so I can track multiple pairs. Or do you know of a way to do this on a large Tv/monitor like this?
Jerry